+++ app/src/main/java/com/example/playlistmaker/library/ui/favourite/FavouriteTracksFragmentState.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.library.ui
+package com.example.playlistmaker.library.ui.favourite
 
 import com.example.playlistmaker.common.domain.model.Track
