+++ app/src/main/java/com/example/playlistmaker/library/ui/favourite/FavouriteTracksViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.library.ui
+package com.example.playlistmaker.library.ui.favourite
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
