+++ app/src/main/res/layout/fragment_favourite_tracks.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".library.ui.FavouriteTracksFragment">
+    tools:context=".library.ui.favourite.FavouriteTracksFragment">
 
     <include
         android:id="@+id/included_rv"
