+++ app/src/main/res/layout/fragment_media_library.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".library.ui.MediaLibraryFragment">
+    tools:context=".library.ui.medialibrary.MediaLibraryFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
